Align example configuration and license installation

+++ benchmarks/lua/lua_capsule_config.h
@@ -1,15 +1,22 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-// Force-included platform hooks for the stock Lua sources.
+// Force-included into stock Lua. Lua deliberately leaves these platform
+// hooks open in luaconf.h/ldo.c, so Capsule does not patch the upstream tree.
 #pragma once
 
-// Work around clang 23's BPF debug-info crash on the parenthesized variadic
-// declarations in Lua's headers by declaring their plain forms first.
+// clang 23 crashes emitting BPF debug info for an extern variadic function
+// whose first declaration is parenthesized, as lua.h and lauxlib.h write
+// them: CGDebugInfo::getOrCreateFunctionType does not look through the type
+// sugar, records no parameters, and the BPF extern-declaration path then
+// indexes past the empty list. Declaring the three plainly first sidesteps
+// it; drop these once the SDK's clang carries the fix.
 struct lua_State;
 extern const char* lua_pushfstring(struct lua_State* L, const char* fmt, ...);
 extern int lua_gc(struct lua_State* L, int what, ...);
 extern int luaL_error(struct lua_State* L, const char* fmt, ...);
 
-void lua_capsule_write(const char* text, unsigned long length);
+#include <stddef.h>
+
+void lua_capsule_write(const char* text, size_t length);
 
 #define lua_writestring(s, l) lua_capsule_write((s), (l))
 #define lua_writeline() lua_capsule_write("\n", 1)
